Add song object for storing playback data
Planning on moving sequencing into its own concern so that it's easier
to "resequence" when a new sample is added/removed.

Adds text formatter for easier "integration" testing and assertion. I
don't think the formatter is in the right spot just yet, but it's been
helpful to get some specs in as things move around.